feat: add course codes to section roomTimes
Before this commit, a section in a timetable would have no way to know which course it belonged to except for the course id. This works for a backend, but for a frontend, for example when rendering the boxes in the grid, to associate it with the right course details, you'd need to do an extra filter.

BREAKING CHANGE: roomTimes now have the course code at the beginning