feat: move authentication tokens to JWTs + fingerprint cookie
See the first few paragraphs of: https://hasura.io/blog/best-practices-of-using-jwt-with-graphql/

BREAKING CHANGE: The session cookie is no longer HTTPOnly, the userInfo cookie is gone, a new fingerprint cookie exists, which is HTTPOnly. The session cookie contains the JWT, and all cookie verifications check if the fingerprint matches the hash in the session cookie.